requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pydantic==2.9.2
python-dotenv==1.0.1
PyJWT==2.9.0
requests==2.32.3
